add viewHolder, adapter , diffutilcallback and small refactoring
