+++ app/src/main/java/com/example/newfilmlistapp/ui/favorites/FavoritesFragment.kt
@@ -4,12 +4,12 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.lifecycle.ViewModelProvider
+import androidx.fragment.app.viewModels
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.newfilmlistapp.databinding.FragmentFavoritesBinding
-import com.example.newfilmlistapp.databinding.FragmentPopularBinding
 import com.example.newfilmlistapp.model.ResultPopular
 import com.example.newfilmlistapp.ui.popular.PopularAdapter
-import com.example.newfilmlistapp.view_model.PopularViewModel
+import com.example.newfilmlistapp.view_model.FavoritesViewModel
 
 
 class FavoritesFragment : androidx.fragment.app.Fragment() {
@@ -18,14 +18,10 @@ class FavoritesFragment : androidx.fragment.app.Fragment() {
     private var totalResults: Int = -1
     private var isLoading: Boolean = false
 
-    private val viewModel: PopularViewModel by lazy {
-        ViewModelProvider(this).get(PopularViewModel::class.java)
-    }
+    private val viewModel: FavoritesViewModel by viewModels()
 
     private val favoriteAdapter: PopularAdapter = PopularAdapter()
     //private val mScrollListener by lazy { RecyclerViewScrollListener(this) }
-
-
 
     private lateinit var binding: FragmentFavoritesBinding
 
@@ -35,18 +31,36 @@ class FavoritesFragment : androidx.fragment.app.Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-
         binding = FragmentFavoritesBinding.inflate(inflater,container,false)
 
+        setFragmentTitle()
+        setRecyclerView()
+
+
+
         return binding.root
+    }
+
+    private fun workWithViewModel() {
+
+        viewModel.favorite.observe(viewLifecycleOwner) {
+
+
+
+
+        }
+
+
+
     }
 
 
     private fun setRecyclerView() {
 
-        binding.recyclerviewFavorites
-
-
+        binding.recyclerviewFavorites.apply {
+            layoutManager = LinearLayoutManager(context)
+            adapter = favoriteAdapter
+        }
 
     }
