+++ app/src/main/java/com/example/newfilmlistapp/ui/favorites/FavoritesViewHolder.kt
@@ -1,0 +1,4 @@
+package com.example.newfilmlistapp.ui.favorites
+
+class FavoritesViewHolder {
+}
