+++ app/src/main/java/com/example/newfilmlistapp/ui/popular/PopularFragment.kt
@@ -106,11 +106,4 @@ class PopularFragment : Fragment() {
         }
     }
 
-
-    private fun toMovieDetail() {
-
-
-    }
-
-
 }
