+++ app/src/main/java/com/example/newfilmlistapp/view_model/MovieDetailViewModel.kt
@@ -21,8 +21,6 @@ class MovieDetailViewModel(): ViewModel() {
 
 
 
-
-
     fun onLoad(movie: MovieDetailWrapper) {
 
         viewModelScope.launch {
